Remove suggestion to use assembly for length check
Replacing the implementation of `_stringEmpty` with

```solidity
uint256 _len;
assembly {
    _len := mload(_s)
}
return _len == 0;
```

works, but does not affect gas costs at all. Experimentation shows that
Solidity doesn't actually copy the underlying buffer when converting a
`string` to a `bytes`, and so consequently `string`s are mutable even in
the assembly-free fragment of Solidity. Today I learned!

The helper itself stays, because IMHO it helps readability.