Use actual EIP-712 typed data signatures
Some manual encoding on the Solidity side, but using the stock Ethers
`_signTypedData` API call on the JavaScript side, so we may have some
confidence that this is to spec.

This is structured in such a way that we can easily switch to a
dual-signature approach that accepts either `\x19Ethereum...` messages
or EIP-712 signatures.