Expose feature trait memberships immediately
Per suggestion of @decentralion, we no longer wait for a feature trait
to be finalized before affirming its memberships. This saves two storage
lookups per `_hasFeatureTrait` call at the cost of exposing partially
constructed states. To ameliorate this downside, we also expose an
accessor to test whether a trait is finalized; frontends should use this
to determine whether to offer users the option to make traitset bids
including a trait.