Stop tracking feature target size
Feature size tracking doesn't interact well with projects like Chromie
Squiggles, whose minting is paused but not completed. We'd like to be
able to initialize feature traits for Squiggles and add members, but we
don't know how many (e.g.) Pipes there will end up being. So, we just
remove this information from the chain state entirely. We still track
and (now) expose the *current* size of feature traits, so that frontends
can check whether the trait is at least up to date with those tokens
that have been minted so far.