implement + test bidder royalties
The bid can now specify a list of royalties, which get paid on-top of
the sale price (vs ask royalties which are deducted from the sale
proceeds).

This is intended to reward agents of the buyer -- e.g. an advisor, or
the frontend that the buyer initiated the transaction from