Cancel order nonces before making external calls
Previously, calls to `weth.transferFrom` and `token.safeTransferFrom`
occurred only at the end of the method. Thus, theoretically, either of
these methods or the ERC-721 receive hook called by `safeTransferFrom`
could attempt to fill the same order again, which shouldn't be allowed.

I can't *quite* see a way that this theoretical attack would be
practically useful. Since we currently only authorize bids via ECDSA
signatures, the bidder must be an externally owned account and not a
contract, and therefore has no ERC-721 receive hook, mitigating the most
obvious source of untrusted code. But nonce reuse is still intended to
be impossible, so we should close this bug all the same.