Emit events on successful trades
Unfortunately, we need to emit two events per trade so that we can index
on the buyer, seller, tokenId, and tradeId.

The tradeId is a deterministic hash of the trade computed based on the
nonce in the bid and ask, and the address of the buyer and seller. Since
nonces are non-reusable, this will be unique per trade (modulo
hypothetical collisions). This saves a lot of gas compared to keeping an
incrementing tradeId in contract storage.