oracle: reduce storage I/O on nearby token IDs
We now only read from and write to storage on crossing a multiple-of-256
boundary in the token ID sequence. It's thus now especially beneficial
for that sequence to be in sorted order, though it's still not required.

wchargin-branch: oracle-reduce-nearby-io
wchargin-source: ac651caa8c05d39f6db4b395bba2ce768b5ec1b2