bidStructHash: array hash is bytes32 not uint256
This fixes an argument to the abiCoder in bidStructHash to properly
record that the hash of an uint256 array is a bytes32, not a uint256.

I added a test to explicitly check that legacy traitset bids work,
although it passes both before and after this change, so I suppose there
was no semantic difference.
(ping @wchargin)