Market: add interface-level support for bundles
This changes the MarketMessages types so that tokenIds is an array, on
both bids and asks, so that we can support bundles.

This does _not_ add implementation level support; instead we revert if
the bid or ask are not specifying exactly one token id. This cleanly
separates the (considerable) complexity of changing the interface, from
the (also considerable) complexity of supporting bundles properly.