market: move contract state into EIP-712 domain
The reason for adding the `Market` address to the EIP-712 domain (as the
`verifyingContract`) was to ensure that signatures for a given token ID
on one NFT project can't be substituted for the same token ID on a
different project (say, Art Blocks #123 vs. Deafbeef #123). This patch
switches the EIP-712 domain from including the market address to
directly including the market state: i.e., the addresses of the ERC-721,
ERC-20, and trait oracle contracts.

This is primarily intended to improve the compatibility story, so that
sibling markets for the same token contract---whether with the same or
different implementation code---can naturally accept these signatures.
However, if my understanding is correct, it also mitigates a theoretical
attack involving reorgs. Consider:

 1. Alice deploys a market to Address X.
 2. Alice initializes the market to Token A.
 3. Bob approves Alice's market to operate his WETH and Token A.
 4. Bob signs and publishes a buy order for Alice's market.
 5. The chain reorgs to before (2).
 6. Alice initializes the market to Token B.

If Bob's signature includes only the market's address X, then his signed
message will appear to authorize a valid order for Token B, which was
never Bob's intent. Moreover, Bob's WETH approval grant in step (3) is
still valid after the reorg and may be replayed on chain. Thus, the
owner of the NFT for Token B corresponding to Bob's original order can
sell that token to Bob, even though Bob only ever intended to place a
buy order for Token A. That's bad.

(Likewise, if the market is reinitialized with a different `weth`
contract, then any sell orders from Bob can be executed with a different
currency, and if the market is reinitialized with a different trait
oracle, then any traitset bids from Bob can be filled with different
tokens than Bob had intended.)

It's worth noting that this doesn't even really require Alice to be
malicious: "Alice" could be a `MarketFactory` smart contract with a
public `createMarket(IERC721 token, ...)` method that uses the `CREATE`
opcode (not `CREATE2`) to create market contracts and then initializes
them in the same transaction. In that case, steps (1--2) and (6) can be
replaced by "Cheryl uses the factory to create a market for Token A" and
"David uses the factory to create a market for Token B", respectively;
if the factory's nonce is rolled back during the reorg, the addresses of
the two markets will be the same. Thus, any "normal" reorg that happens
to span a market creation and approvals to the newly created market may
be a target for this attack.

If Bob's signature includes the address of Token A instead of the
address of Alice's market, then there is no such confusion. The
signature now describes Bob's intent explicitly, instead of relying on
"the state of `verifyingContract` at *the current head* of `chainId`" to
resolve the meaning. Once Alice initializes the market to Token B in
step (6), the market will refuse to honor any signatures whose salts
describe Token A, and likewise for the other variables, as desired.