market: check authorization just before transfer
This patches a time-of-check/time-of-use (TOCTOU) attack that took
advantage of the bundle transfer implementation. Previously, all
authorization checks were performed upfront, before any transfers
occurred, but performing a transfer requires calling out to external
code (notably, ERC-721 receive hooks) that could invalidate those
authorizations.

A pernicious version of the attack in question works as follows:

 1. Alice and Bob both use a certain market and have given it approval
    to manage all their tokens.
 2. Alice owns `token1` and `token2`, stored in a smart contract with a
    custom ERC-721 receive hook.
 3. Bob publishes a bid order for `token2`.
 4. Alice creates an ask for the bundle containing `[token1, token2]` in
    that order, and approves it from her contract account. She also
    creates and approves a corresponding bid for the same bundle.
 5. Alice executes the wash sale that she just approved. The market
    transfers `token1` from Alice to Alice, executing the ERC-721
    receive hook on her contract.
 6. Alice's ERC-721 receive hook approves an ask corresponding to the
    bid on `token2` that Bob made in step (2).
 7. Still in the receive hook, Alice fills the order for `token2`. Bob's
    WETH is transferred to Alice, and `token2` is transferred to Bob.
 8. The receive hook finishes.
 9. Having finished the transfer of `token1`, the market continues
    executing the bundle sale, by transferring `token2` from its current
    owner, Bob, to the bidder, Alice.

The final state is that both `token1` and `token2` still belong to
Alice, but Bob's WETH has been sent to Alice, and he has nothing to show
for it. That's bad.

(Note that the sale of `token2` in steps (6) and (7) could happen in a
different Archipelago market or outside of the Archipelago system
altogether. Re-entrancy guards won't save you here.)

By moving the authorization checks to immediately before the transfer,
the market notices that `token2` is no longer owned by the asker, and
therefore rejects the trade.

This patch includes a contract that actually executes this attack.
Before the change to `Market.sol`, the new test case fails with
"successfully executed TOCTOU attack". After the change, the test
passes.

I haven't carefully reviewed the rest of the code to see if there are
similar attack surfaces; we should do so.

wchargin-branch: market-owner-toctou
wchargin-source: 49144c7f77607e85333640737db97c35e7ca7e5b