market: add `verifyingContract` to domain separator
In e1916b3bd179, we removed the `verifyingContract` field from the
domain separator to allow sibling markets for the same token contract to
share a signing domain. But since the cancellation state is recorded
only in the contract storage of the market itself, approving a sibling
market (or a hypothetical "V2" market) to manage your tokens also
effectively entails "un-cancelling" all previous orders. That's bad.

We considered factoring out a separate "cancellation store" contract
that could be shared by multiple markets, which can be managed, but
introduces additional complexity and also (in our prevailing proposal)
requires the user to grant upfront approval that the market may manage
their cancellation state, which is a bit annoying. We think that this
flavor of liquidity fragmentation is not too big of a deal, especially
given that a frontend can paper over it by routing orders appropriately.

I still like including the addresses of the ERC-721, ERC-20, and token
oracle contracts in the salt, since it provides assurance about those
aspects of the signing domain without relying on the current chain state
and also avoids any reorg concerns.