Market: Avoid potential TOCTOU attacks
This re-applies the TOCTOU mitigation from #5 in the non-bundle case.
Essentially, we check that the asker is the token's owner or approved
operator immediately before transferring it to the recipient. This way,
even if there were a malicious contract operation during the method
(e.g. an evil WETH contract), we won't transfer the token if it doesn't
properly belong to the seller.