Bids and asks include the token address
We're planning to refactor so that Archipelago will have a single
(cross-contract) Market contract, rather than a Market-per-ERC721.

To facilitate this, we need to put the ERC721 token address into the
Bid/Ask data types. Since this is a finnicky change, I'm doing this as
its own commit prior to actually using this information in order
resolution. (I.e. as of this commit, its vestigial, or perhaps
"pre-stigial".)

Test plan: `npm test` passes.