Make Market contract generic across ERC721s
This commit upgrades the Market contract so that it now gets the token
address from the bid and ask, rather than requiring that the token
address be hardcoded into the Market at initialization time. This means
we can deploy a single Market for all of Archipelago, rather than
needing one-per-contract.

Test plan: `npm test` passes; also added a new test verifying that the
bid and ask need to actually agree on the token address as well as the
tokenId (or traits).