market: get traitOracle from the bid
This refactors the market code so that it uses the bidder's provided
traitOracle, rather than the one that was initialized on the Market.
To distinguish that the new codepath is definitely working, I've changed
the test code so it provides the zero address as the "canonical" oracle
for the market as a whole.