market: actually use currency from the bid/ask
We now explicitly admit the possibility that the currency won't be weth;
as such, we now test that filling orders with raw ETH, or unwrapping
proceeds to ETH, will fail if the currency is not weth.

Also, add a test to ensure that the bid and ask actually agree about the
currency.