market: remove initialization and salt
Now that we don't need a hardcoded weth value on the market contract
(because bids/asks specify their currencies), we can remove the
initialization logic and the salt entirely. This makes the contract
pleasingly general-purpose, there will now only be a single
ArchipelagoMarket rather than many separate markets for different
contracts.

We no longer prevent users from mistakenly sending ETH to this contract,
since we don't know the canonical address for the weth contract. We
should add an admin-only skim function so we can do something with ETH
that mistakenly gets sent here.