re-order Bid/Ask fields for consistency
Bids and asks have a lot of structural similarities. While it doesn't
make sense to reify an explicit shared Order object (for performance
reasons), we can re-organize the structs and comments to make the shared
structure easy to uncover.

Since I changed the struct field ordering, I also changed the hash field
ordering, per @wchargin's suggestion. This was just one swap (on the Bid
side) -- the change to the Ask struct field ordering actually brought it into
consistency with the hash ordering.