Pass NFT contract address to trait oracle
Given that the Market is no longer scoped to individual NFTs, it's odd
if we retain such scoping at the Oracle level. Since the new field is
not needed for the one "real" oracle implementation (ie the ArtBlocks
oracle), it's mostly just a plumbing change. The ArtBlocks oracle simply
ignores the new variable.