Remove timestamp based cancellations
This is a material gas saving for every single order filled, and the
behavior of mass cancelling outstanding orders can already be
accomplished via a bulk nonce cancellation. It makes more sense to have
the (rare) users of mass cancellations pay the required gas, rather than
imposing it on everyone who uses the market.

Saves over 5k gas on the common single-token fill.