Add mechanism for market emergency shutdowns
This will allow us to deactivate the market in the event we discover a
critical security vulnerability.

In the past, I advocated for creating a builtin mechanism for
auto-destroying the shutdown mechanism (e.g. 12 months after deploy), to
prevent abuse or coercion by the contract owner. Currently, I'm feeling
like it's a bit of unnecessary complexity; there's not much upside for
the owner in shutting down the marketplace, and if we want to remove
that risk, we can always renounce ownership of the contract. However,
that would remove the ability to skim ETH that accumulates at the market
contract. My sense is that this is fine; would be open to creating a
separate "shutdown the shutdown mechanism" call if we want to be able to
revoke that without renouncing ownership, but atm it's feeling like
unnecessary complexity.