market: include a hardcoded protocol royalty
This royalty is enforced at the protocol level, and allows us to
guarantee that the protocol treasury receives revenue even if some
markets misbehave.

Updated the gas script to include this royalty. This makes our gas
numbers look worse, particularly in the "standard fill (3 royalties)"
case; however, that case was inappropriately not paying separate
protocol + frontend royalties, so it actually wasn't realistic
previously.