circuit: test underlying oracles returning non-bools
Feels appropriate to be a bit defensive here: since we reinterpret-cast
the result of `hasTrait(...)` directly to a `uint256` and then use that
as a mask, it'd be problematic if the return word could be any value
either than `0` or `1`. This test just makes sure that Solidity properly
checks that for us.