artblocks: keep hash chain of updates for each trait
This way, we can safely create finalization messages of the form
"finalize the first `k` tokens for trait `t`, where the update chain
hash is `h`" without worrying about this signed message being replayed
on an earlier fork of the chain where not all the tokens had yet been
finalized.

This does require an extra storage cell per trait. Maybe I'll change
this to pack it in with the size of each trait: trait sizes are bounded
by 1 million (20 bits), and we can tolerate dropping half the entropy of
the hash chain. Once finalization states are changed to "all up to..."
indices, we can write that into the same word, too; something like:

```solidity
struct TraitState {
    /// Number of tokens enrolled in this trait.
    uint32 count;
    /// Largest `k` such that the first `k` tokens are finalized.
    uint32 finalized;
    /// Cumulative hash of all updates to this trait (truncated).
    bytes16 updateHash;
}
```

Then, even with this change, we'll be net saving storage compared to
before it.