artblocks: finalize prefixes, not words
Each trait now tracks a single number `_k` to indicate that the first
`_k` tokens for the trait have been finalized, instead of tracking
finalization all-or-nothing at the word level. This means that we can
finalize a token's membership before all the tokens in its word are
minted, and it also means that we have less data to store: only one
scalar bounded by 1 million, which can be packed into existing storage
cells, as opposed to one bit for every word of the trait, which is
wasteful for most traits.

It's no longer possible to finalize a token without finalizing all the
tokens before it, but that should be fine, since tokens are always
minted in order, by definition.