artblocks: add `tokenContract` to info messages
Not yet written, but will be used to ensure that we only attest traits
for tokens on the right contract.

This isn't a security issue currently because the market will only fill
a order for a bid and an ask that explicitly agree on the token
contract. However, it's more correct to only attest traits for tokens on
the right contract, and it's better defense in depth.