Refactor orders around shared "Agreements"
Now the Bid and Ask both have the hash of an Agreement, which is passed
separately to the market. This allows us to deduplicate shared fields
between the Bid and Ask and simplify the logic around checking that they
agree on shared fields. Saves 5996 gas on a "standard fill" with three
required royalties plus the protocol royalty.

Paired with @wchargin