circuit: remove use of word-unaligned bytestrings
Previously, we decoded the trait data by consuming the `_buf` bytestring
as we read it. This worked, and avoided some superfluous memory copies
when transforming buffers, and the approach is entirely sensible at the
EVM level---but I've been unable to find explicit endorsement of using
such word-unaligned bytestrings within Solidity, or any guarantees that
the compiler (perhaps in future versions) won't emit code that
implicitly assumes that `bytes` values are word-aligned. To be
conservative, this patch removes all manual bytestring manipulation.

Naive implementations of this change dramatically increase the gas
needed. However, with this patch, we can also keep `_buf` entirely in
calldata instead of copying it out to memory, and we can use Solidity's
slice values directly at the external `call` site to work around some
additional memory copies. In principle, Solidity should be able to
compile this efficiently, directly copying a slice of the buffer from
calldata into the region of memory used to build the outgoing calldata
*and then freeing that entire region* once the call completes.
Impressively, this patch actually saves 94 gas on net for a typical fill
(single union/intersection of two traits).