market: authorize orders from `msg.sender` via `EXTERNAL` auth
This way, if one side of an order has been signed, a counterparty can
fill the order by only signing and submitting a `fillOrder` transaction,
rather than needing to separately sign their side of the order first.
This usability improvement is especially notable for signing flows that
involve hardware or cold storage wallets.

To use this flow, set the signature kind to `EXTERNAL` and set the
signature to `abi.encode(yourAddress)`, where `yourAddress` is the
address of the direct caller of `fillOrder`, which may be either an
externally owned account or a contract.