market: emit nonce cancellation events on fills
Previously, explicitly calling `cancelNonces` would emit an event, but
calling `fillOrder` would cancel the nonces for the bid and the ask
without emitting any events to that effect. The only events emitted were
`Trade`, `TokenTraded`, and `RoyaltyPaid`, neither of which contains
enough information to determine which nonces were cancelled. (The nonces
flow into the trade ID, but only before a Keccak hash, so it's not
possible to recover the nonces from the trade ID alone.)

With this patch, it's now possible to determine from the event log alone
which nonces are cancelled.

This costs an extra 3081 gas per fill: 3k gas due to the two `LOG3`
instructions, 81 gas due to data manipulation overhead.